+++ packages/app/admin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fougere/admin
 
+## 0.8.3-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @fougere/core@0.8.3-alpha.0
+  - @fougere/schema@0.8.3-alpha.0
+
 ## 0.8.2-alpha.0
 
 ### Patch Changes